% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhdf5_api.R
\name{deleteDomain}
\alias{deleteDomain}
\title{Function to delete a file on hdf server}
\usage{
deleteDomain(url, domain)
}
\arguments{
\item{url}{character string with http server url}

\item{domain}{character string with domain name to be created}
}
\value{
r http response object
}
\description{
Function to delete a file on hdf server
}
\note{
\url{"http://170.223.248.164:7248"}
}
\examples{
if (nchar(Sys.getenv("password"))>0) {
  tstring = sub("\\\\/", "", tempfile(tmpdir=""))
  dom = deleteDomain("http://170.223.248.164:7248", paste(tstring, ".hdfgroup.org", sep=""))
  readBin(dom$content, what="character")
  }
}
