% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhdf5_api.R
\name{postDataset}
\alias{postDataset}
\title{Function to create a dataset in a file with extensible dimensions}
\usage{
postDataset(url, domain, type, shape, maxdims)
}
\arguments{
\item{url}{character string with http server url}

\item{domain}{character string with domain name to be created}

\item{type}{character string with dataset type like H5T_IEEE_F32LE}

\item{shape}{numeric vector with initial dimensions of dataset}

\item{maxdims}{numeric vector with maximum extent of each dimension or 0 for unlimited dimension}
}
\value{
r http response object
}
\description{
Function to create a dataset in a file with extensible dimensions
}
\note{
\url{"http://170.223.248.164:7248/datasets"}
}
\examples{
if (nchar(Sys.getenv("password"))>0) {
    tstring = sub("\\\\/", "", tempfile(tmpdir=""))
    dom = putDomain("http://170.223.248.164:7248", paste(tstring, ".hdfgroup.org", sep=""))
    ds = postDataset(url="http://170.223.248.164:7248/datasets",
        domain=paste(tstring, ".hdfgroup.org", sep=""),  
        type="H5T_IEEE_F32LE",
        shape=c(10,5), maxdims=c(0,5))
    ans = readBin(ds$content, what="character")
    ans
    }
}
