% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhdf5_api.R
\name{putValue}
\alias{putValue}
\title{Function to insert values into dataset}
\usage{
putValue(url, domain, value, start, stop)
}
\arguments{
\item{url}{character string with http server url}

\item{domain}{character string with domain name to be created}

\item{value}{list with values to be inserted into dataset}

\item{start}{(optional)integer/integer array with starting coordinate of selection to be updated}

\item{stop}{(optional)integer/integer array with ending coordinate of selection to be updated}
}
\value{
r http response object
}
\description{
Function to insert values into dataset
}
\note{
\url{"http://170.223.248.164:7248/datasets/dsetuuid/value"}
}
